// types/components.ts
export type MusicCardItem = {
  id: string;
  title: string;
  subtitle: string;
  image: string;
};

export type ProfileCardItem = {
  id: string;
  name: string;
  songTitle: string;
  image: string;
};

export type RecommendedArtistItem = {
  id: string;
  name: string;
  tag: string;
  image: string;
};